Mar25 fixes to support Python3.12 - increased version to 0.4.0 - moved nodestate, heartbeat from ros_tools into robot_pack.py as they depend on aarapsi_robot_pack - added try/except around importing robot_pack.py as aarapsi_robot_pack should not be compulsory - added weights_only=False for torch.load calls, as torch.load default changes from False to True - fix: nn.nn_factors 'is False' to '== False' for newer versions of numpy - nn.params now tries to make specified directories - DataLoadSaveError in VPRDatasetProcessor.generate_upd_dataset now provides params which fail in error output - vpr.classes.download_models data now ships with installation - vpr.classes.download_models now actually behaves properly with some more friendly outputs - vpr.classes.download_models SALAD data link alive again - add: vpr.classes.datahelpers, as I used this to debug a few things and they're kinda neat - remove duplicate destroy call in apgem.py's APGEMContainer.__del__ for self.whitenv
